Monster Class

+++ src/main/java/org/example/model/ProductForSale.java
@@ -2,14 +2,14 @@ package org.example.model;
 
 public abstract class ProductForSale {
     private String type;
-    private double price;
+    private int price;
     private String description;
 
     public String getType() {
         return type;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
@@ -17,7 +17,7 @@ public abstract class ProductForSale {
         return description;
     }
 
-    public ProductForSale(String type, double price, String description) {
+    public ProductForSale(String type, int price, String description) {
         this.type = type;
         this.price = price;
         this.description = description;
